+++ src/test/java/amidst/mojangapi/mocking/FakeWorldBuilder.java
@@ -14,7 +14,7 @@ import amidst.mojangapi.world.testworld.TestWorldDeclaration;
 import amidst.mojangapi.world.testworld.TestWorldEntryNames;
 import amidst.mojangapi.world.testworld.file.TestWorldDirectory;
 import amidst.mojangapi.world.testworld.file.TestWorldDirectoryDeclaration;
-import amidst.mojangapi.world.testworld.storage.BiomeDataJson;
+import amidst.mojangapi.world.testworld.storage.json.BiomeDataJson;
 import amidst.mojangapi.world.testworld.storage.json.WorldMetadataJson;
 
 @ThreadSafe
