+++ src/test/java/amidst/mojangapi/world/testworld/storage/json/BiomeDataJson.java
@@ -1,4 +1,4 @@
-package amidst.mojangapi.world.testworld.storage;
+package amidst.mojangapi.world.testworld.storage.json;
 
 import java.util.Map;
